- Changed physical addresses from uintptr_t to uint64_t - Replaced usage of void for marking as unused by using __unused - Adjusted virtual memory management to work with uint64_t as physical address - Changed context within virt context type to uint64_t

+++ include/arch/arm/v6/mm/virt/short.h
@@ -23,10 +23,10 @@
 
 #include <kernel/mm/virt.h>
 
-void v6_short_map( virt_context_ptr_t, uintptr_t, uintptr_t, uint32_t );
+void v6_short_map( virt_context_ptr_t, uintptr_t, uint64_t, uint32_t );
 void v6_short_map_random( virt_context_ptr_t, uintptr_t, uint32_t );
 void v6_short_unmap( virt_context_ptr_t, uintptr_t );
-uintptr_t v6_short_create_table( virt_context_ptr_t, uintptr_t, uintptr_t );
+uint64_t v6_short_create_table( virt_context_ptr_t, uintptr_t, uint64_t );
 void v6_short_set_context( virt_context_ptr_t );
 void v6_short_prepare_temporary( virt_context_ptr_t );
 virt_context_ptr_t v6_short_create_context( virt_context_type_t );
