+++ include/arch/arm/v7/mm/virt/short.h
@@ -23,10 +23,10 @@
 
 #include <kernel/mm/virt.h>
 
-void v7_short_map( virt_context_ptr_t, uintptr_t, uintptr_t, uint32_t );
+void v7_short_map( virt_context_ptr_t, uintptr_t, uint64_t, uint32_t );
 void v7_short_map_random( virt_context_ptr_t, uintptr_t, uint32_t );
 void v7_short_unmap( virt_context_ptr_t, uintptr_t );
-uintptr_t v7_short_create_table( virt_context_ptr_t, uintptr_t, uintptr_t );
+uint64_t v7_short_create_table( virt_context_ptr_t, uintptr_t, uint64_t );
 void v7_short_set_context( virt_context_ptr_t );
 void v7_short_prepare_temporary( virt_context_ptr_t );
 virt_context_ptr_t v7_short_create_context( virt_context_type_t );
